+++ src/app/locales/en/translations.js
@@ -36,13 +36,17 @@ import archiveRecall from './components/archive-recall';
 import fileRecall from './components/file-recall';
 import archiveContentCount from './components/archive-content-count';
 import archiveState from './components/archive-state';
-import archivePropertiesModal from './components/archive-properties-modal';
+import archiveDetailsModal from './components/archive-details-modal';
 import fileInfoModal from './components/file-info-modal';
 import fileMetadata from './components/file-metadata';
 import filePermissions from './components/file-permissions';
 import fileShares from './components/file-shares';
-import fileQos from './components/file-qos';
-import userInfoPopover from './components/user-info-popover';
+import fileQos from './components/file-qos'; <<
+<< << < HEAD
+import userInfoPopover from './components/user-info-popover'; ===
+=== =
+import archiveAuditLog from './components/archive-audit-log'; >>>
+>>> > 8215587 ae8c2128c0b9dcc0a43d38a0295c69e3a
 
 import uploadManager from './services/upload-manager';
 
@@ -64,6 +68,7 @@ import fileMetadataViewModel from './utils/file-metadata-view-model';
 import filePermissionsViewModel from './utils/file-permissions-view-model';
 import fileQosViewModel from './utils/file-qos-view-model';
 import fileDistributionViewModel from './utils/file-distribution-view-model';
+import archivePropertiesViewModel from './utils/archive-properties-view-model';
 
 const translations = {
   components: {
@@ -102,13 +107,14 @@ const translations = {
     fileRecall,
     archiveContentCount,
     archiveState,
-    archivePropertiesModal,
+    archiveDetailsModal,
     fileInfoModal,
     fileMetadata,
     filePermissions,
     fileShares,
     fileQos,
     userInfoPopover,
+    archiveAuditLog,
   },
   services: {
     uploadManager,
@@ -129,6 +135,7 @@ const translations = {
     fileQosViewModel,
     fileDistributionViewModel,
     fileInfo,
+    archivePropertiesViewModel,
     archiveForm: {
       baseModel: archiveFormBaseModel,
       viewModel: archiveFormViewModel,