+++ src/app/locales/en/utils/archive-properties-view-model.js
@@ -1,0 +1,3 @@
+export default {
+  modifyingArchive: 'modifying archive properties',
+};